common-file-formats/requirements.txt
appnope==0.1.2
astroid==2.3.3
autopep8==1.5.7
backcall==0.2.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==5.5
decorator==5.0.9
dependency-injector==4.32.3
docopt==0.6.2
Flask==1.1.1
grip==4.5.2
idna==2.8
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.3
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
path-and-address==2.0.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.22.0
six==1.13.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.25.7
wcwidth==0.2.5
Werkzeug==0.16.0
wrapt==1.11.2
